Create a mockUser generator and move mock generators into folder
