+++ test-utils/mockGenerators/mockPassword.js
@@ -1,0 +1,15 @@
+import faker from 'faker';
+import { minPasswordCharNum } from '../../common/constants/validations';
+
+/**
+ * @description generate a mocked password for registration tests
+ *
+ * @export
+ * @param {boolean} [isValid=true] Determine if the returned password is valid
+ * @returns {string}
+ */
+export default function mockPassword(isValid = true) {
+  return isValid
+    ? `${faker.internet.password(minPasswordCharNum)}!1Aa`
+    : `${faker.internet.password(minPasswordCharNum - 2)}`;
+}
